+++ relational-database/crud_majors.py
@@ -43,7 +43,7 @@ def main():
     # Get the user's menu choice.
     while choice != menu_exit:
         choice = get_menu_choice()
-        execute_choice(choice, cur)
+        execute_choice(cur, choice)
 
     # Commit the changes.
     conn.commit()
@@ -76,7 +76,7 @@ def get_menu_choice():
 
 
 # Perform the action that the user selected.
-def execute_choice(choice, cur):
+def execute_choice(cur, choice=0):
     if choice == 1:  # Add a new major
         create_new_major(cur)
     elif choice == 2:  # Search for an existing major
@@ -94,15 +94,14 @@ def create_new_major(cur):
     print("Please add New Major Information")
     print("")
     new_name = input('Please enter the Name of the new Major: ')
-    
-    cur.execute('''INSERT INTO Majors (Name) 
+    new_name = new_name.strip()
+    if new_name != "":
+        cur.execute('''INSERT INTO Majors (Name) 
                     VALUES (?)''', (new_name,))
     
-    # Add new record verification here
-    # select new_name
-    # if found - print Added, else print not added
-    cur.execute('''SELECT * FROM Majors WHERE Name = (?)''', (new_name,))
-    display_majors(cur, 1)
+        # if found - print Added, else print not added
+        cur.execute('''SELECT * FROM Majors WHERE Name = (?)''', (new_name,))
+        display_majors(cur, 1)
 
     
 # the find_major function searches for a specific major in the Majors table
@@ -111,9 +110,9 @@ def find_major(cur):
     print("Please enter the Major you are looking for")
     print("")
     name = "%" + input('Please enter the Name of the Major to be found: ') + "%"
-    
-    cur.execute('''SELECT * FROM Majors WHERE name LIKE (?)''', (name,))
-    display_majors(cur, 2)
+    if name != "":
+        cur.execute('''SELECT * FROM Majors WHERE name LIKE (?)''', (name,))
+        display_majors(cur, 2)
 
 
 def update_major(cur):
@@ -122,11 +121,14 @@ def update_major(cur):
     display_majors(cur, 5)
     print("")
     select_id = input("Please enter the MajorID number of the Major you wish to change: ")
-    new_name = input("Please enter the new name for the Major: ")
-
-    cur.execute('''UPDATE Majors SET name = (?) WHERE MajorID = (?)''', (new_name, select_id))
-    cur.execute('''SELECT * FROM Majors WHERE MajorID = (?)''', (select_id,))
-    display_majors(cur, 3)
+    select_id = select_id.strip()
+    if select_id != "":
+        new_name = input("Please enter the new name for the Major: ")
+        new_name = new_name.strip()
+        if new_name != "":
+            cur.execute('''UPDATE Majors SET name = (?) WHERE MajorID = (?)''', (new_name, select_id))
+            cur.execute('''SELECT * FROM Majors WHERE MajorID = (?)''', (select_id,))
+            display_majors(cur, 3)
 
 def delete_major(cur):
     print("")
@@ -134,11 +136,27 @@ def delete_major(cur):
     print("Which of the following Majors do you wish to delete?")
     display_majors(cur, 5)
 
-    select_id = input("Please enter the ID number of the record you wish to delete: ")
+    verify_choice = 0
+    while verify_choice == 0:
+        select_id = input("Please enter the ID number of the record you wish to delete: ")
+        select_id = select_id.strip()
+        if select_id != "":
+            # If select_id is being used in Students Table, message that it cannot be deleted as it is in use
+            cur.execute('''SELECT MajorID FROM Students WHERE DeptID = (?)''', (select_id,))
+            results = cur.fetchone
+            
+            # If MajorID is not used in Students Table, Okay to delete
+            if results == None: 
+                cur.execute('''DELETE FROM Majors WHERE MajorID = (?)''', (select_id))
+                cur.execute('''SELECT * FROM Majors WHERE MajorID = (?)''', (select_id,))
+                display_majors(cur, 4)
+                verify_choice = 1
+            # If MajorID is used in Students Table, message that it cannot be deleted 
+            else:
+                print(f'\033[1mMajor ID cannot be deleted because it is in use in the Students Table.\033[0m')
+        else:
+            verify_choice=1
 
-    cur.execute('''DELETE FROM Majors WHERE MajorID = (?)''', (select_id))
-    cur.execute('''SELECT * FROM Majors WHERE MajorID = (?)''', (select_id,))
-    display_majors(cur, 4)
 
 # The display_majors function displays the contents of the Majors table.
 def display_majors(cur, value=5):
@@ -170,7 +188,19 @@ def display_majors(cur, value=5):
         name = row[1]
         print(f'{major_id:^9} {name:<20}')
 
+# Verify that the value is a valid MajorID
+def verify_major(cur, value):
+    cur.execute('''SELECT * FROM Majors WHERE MajorID = (?)''', (value,))
+    results = cur.fetchone()
+    print(f"value {value}, {results}, {results[0]} - {results[1]}")
+    if results == None:
+        return False
+    else:
+        return True
 
+
+        
+    
 # Call the main function ONLY if the file is being run as a standalone program.
 if __name__ == "__main__":
     main()
